requirements.txt
flask
flask-cors
streamlit
faiss-cpu
sentence-transformers
pypdf
pdfplumber
python-docx
google-genai
numpy
python-dotenv
